+++ tests/test_class.py
@@ -9,10 +9,11 @@ import google.cloud.exceptions
 import pytest
 from django.core.exceptions import SuspiciousFileOperation
 from django.utils import six
-from django.utils.crypto import get_random_string
 
 from django_gcloud_storage import safe_join, remove_prefix, GCloudFile
 
+from conftest import TEST_FILE_CONTENT, TEST_FILE_PATH
+from helpers import upload_test_file
 
 def urlopen(*args, **kwargs):
     try:
@@ -27,16 +28,6 @@ def urlopen(*args, **kwargs):
         pass
 
     return urlopen(*args, **kwargs)
-
-
-@pytest.fixture
-def storage(storage_object):
-    # Make sure there are no test files due to a previous test run
-    from django_gcloud_storage import prepare_name
-    for blob in storage_object.bucket.list_blobs():
-        storage_object.bucket.delete_blob(prepare_name(blob.name))
-
-    return storage_object
 
 
 # noinspection PyClassHasNoInit,PyMethodMayBeStatic
@@ -125,25 +116,8 @@ class TestGCloudFile:
 
 # noinspection PyClassHasNoInit,PyMethodMayBeStatic
 class TestGCloudStorageClass:
-    TEST_FILE_NAME = "test_file_" + get_random_string(6)
-    TEST_FILE_NAME_UNICODE = "test_file_陰陽_" + get_random_string(6)
-    TEST_FILE_CONTENT = "Brathähnchen".encode("utf8")
-
-    def upload_test_file(self, storage, name, content):
-        if six.PY3 and isinstance(content, str):
-            content = content.encode("utf8")
-
-        with TemporaryFile() as testfile:
-            testfile.write(content)
-            testfile.seek(0)
-            storage.save(name, testfile)
-
-    def test_should_create_blob_at_correct_path(self, storage):
-        with TemporaryFile() as testfile:
-            testfile.write(self.TEST_FILE_CONTENT)
-            testfile.seek(0)
-            storage_path = storage.save(self.TEST_FILE_NAME, testfile)
-        assert storage_path == self.TEST_FILE_NAME
+    def test_should_create_blob_at_correct_path(self, storage, test_file):
+        assert test_file == TEST_FILE_PATH
 
     def test_should_create_a_valid_client_object(self, storage):
         with pytest.raises(google.cloud.exceptions.NotFound):
@@ -152,109 +126,96 @@ class TestGCloudStorageClass:
     def test_should_create_a_valid_bucket_object(self, storage):
         assert storage.bucket.exists()
 
-    def test_should_be_able_to_save_and_open_files(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
+    def test_should_be_able_to_save_and_open_files(self, storage, test_file):
+        f = storage.open(test_file)
+        assert f.read() == TEST_FILE_CONTENT
 
-        f = storage.open(self.TEST_FILE_NAME)
-        assert f.read() == self.TEST_FILE_CONTENT
+    def test_should_return_created_time(self, storage, test_file):
+        assert isinstance(storage.created_time(test_file), datetime.datetime)
 
-    def test_should_return_created_time(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
-
-        assert isinstance(storage.created_time(self.TEST_FILE_NAME), datetime.datetime)
-
-    def test_should_return_modified_time(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
-
-        assert isinstance(storage.modified_time(self.TEST_FILE_NAME), datetime.datetime)
-        assert isinstance(storage.get_modified_time(self.TEST_FILE_NAME), datetime.datetime)
+    def test_should_return_modified_time(self, storage, test_file):
+        assert isinstance(storage.modified_time(test_file), datetime.datetime)
+        assert isinstance(storage.get_modified_time(test_file), datetime.datetime)
 
     def test_should_be_able_to_delete_files(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
-        storage.delete(self.TEST_FILE_NAME)
+        file_name = "test_delete_file"
+        upload_test_file(storage, file_name, "")
+        storage.delete(file_name)
 
         # Should not raise an exception by google.cloud
         assert storage.delete("missing_file") is None
 
-    def test_exists_method(self, storage):
-        assert not storage.exists(self.TEST_FILE_NAME)
+    def test_exists_method(self, storage, test_file):
+        assert not storage.exists("missing_file")
+        assert storage.exists(test_file)
 
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
+    def test_should_return_correct_file_size(self, storage, test_file):
+        assert storage.size(test_file) == len(TEST_FILE_CONTENT)
 
-        assert storage.exists(self.TEST_FILE_NAME)
-
-    def test_should_return_correct_file_size(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
-        assert storage.size(self.TEST_FILE_NAME) == len(self.TEST_FILE_CONTENT)
-
-    def test_should_return_publicly_downloadable_url(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
-        assert urlopen(storage.url(self.TEST_FILE_NAME)).read() == self.TEST_FILE_CONTENT
+    def test_should_return_publicly_downloadable_url(self, storage, test_file):
+        assert urlopen(storage.url(test_file)).read() == TEST_FILE_CONTENT
 
     def test_should_work_with_utf8(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME_UNICODE, self.TEST_FILE_CONTENT)
+        file_name = "test_utf8_file_陰陽"
+        upload_test_file(storage, file_name, "")
 
-        storage.exists(self.TEST_FILE_NAME_UNICODE)
+        storage.exists(file_name)
 
         # Don't explode when trying to find a available name for existing files...
-        self.upload_test_file(storage, self.TEST_FILE_NAME_UNICODE, self.TEST_FILE_CONTENT)
+        upload_test_file(storage, file_name, "")
 
     def test_should_be_able_to_list_dirs_and_files(self, storage):
+        file_name = "test_listdir_file"
         subdir_file_pattern = "/subdir/%s.%d"
 
         for i in range(1, 11):
-            self.upload_test_file(storage, subdir_file_pattern % (self.TEST_FILE_NAME, i), "")
+            upload_test_file(storage, subdir_file_pattern % (file_name, i), "")
 
-        self.upload_test_file(storage, "/subdir/a/" + self.TEST_FILE_NAME, "")
-        self.upload_test_file(storage, "/subdir/b/" + self.TEST_FILE_NAME, "")
+        upload_test_file(storage, "/subdir/a/" + file_name, "")
+        upload_test_file(storage, "/subdir/b/" + file_name, "")
 
         # Make sure paths prefixed with a slash are normalized
         assert storage.listdir("") == storage.listdir("/")
         assert storage.listdir("subdir") == storage.listdir("/subdir")
 
-        root_list_dir = storage.listdir("")
-        assert len(root_list_dir[0]) == 1 and len(root_list_dir[1]) == 0
-        assert root_list_dir[0] == ["subdir"]
-
         subdir_list_dir = storage.listdir("subdir/")
         assert len(subdir_list_dir[0]) == 2 and len(subdir_list_dir[1]) == 10
         assert subdir_list_dir[0] == ["a", "b"]
-        assert subdir_list_dir[1][0] == "%s.%d" % (self.TEST_FILE_NAME, 1)
+        assert subdir_list_dir[1][0] == "%s.%d" % (file_name, 1)
 
-    def test_should_not_overwrite_files_on_save(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, "")
-        self.upload_test_file(storage, self.TEST_FILE_NAME, "")
-
-        assert len(storage.listdir("")[1]) == 2
+    def test_should_not_overwrite_files_on_save(self, storage, test_file):
+        duplicate_file = upload_test_file(storage, test_file, "")
+        assert duplicate_file != test_file
 
     def test_changed_files_should_be_reuploaded(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, "")
-        first_modified_time = storage.modified_time(self.TEST_FILE_NAME)
-        local_tmpfile = storage.open(self.TEST_FILE_NAME)
+        file_name = "test_changed_file_reuploaded"
+        file_content = "gcloud".encode("ascii")
+
+        upload_test_file(storage, file_name, "")
+        first_modified_time = storage.modified_time(file_name)
+        local_tmpfile = storage.open(file_name)
 
         assert local_tmpfile.read() == "".encode("ascii")
         local_tmpfile.seek(0)
 
-        local_tmpfile.write(self.TEST_FILE_CONTENT)
+        local_tmpfile.write(file_content)
         local_tmpfile.close()
 
-        assert storage.open(self.TEST_FILE_NAME).read() == self.TEST_FILE_CONTENT
-        assert storage.modified_time(self.TEST_FILE_NAME) != first_modified_time
+        assert storage.open(file_name).read() == file_content
+        assert storage.modified_time(file_name) != first_modified_time
 
     def test_open_should_be_able_to_create_new_file(self, storage):
-        content = self.TEST_FILE_CONTENT
-        if six.PY3 and isinstance(content, str):
-            content = content.encode("utf8")
+        file_name = "test_open_creates_file"
+        file_content = "".encode("ascii")
 
-        with storage.open(self.TEST_FILE_NAME, mode="wb") as f:
-            f.write(content)
+        with storage.open(file_name, mode="wb") as f:
+            f.write(file_content)
 
-        assert storage.exists(self.TEST_FILE_NAME)
+        assert storage.exists(file_name)
 
-    def test_use_unsigned_urls_option(self, storage):
-        self.upload_test_file(storage, self.TEST_FILE_NAME, self.TEST_FILE_CONTENT)
+    def test_use_unsigned_urls_option(self, storage, test_file):
         storage.use_unsigned_urls = True
-        url = storage.url(self.TEST_FILE_NAME)
+        url = storage.url(test_file)
         storage.use_unsigned_urls = False
         for i in ["Signature=", "GoogleAccessId=", "Expires="]:
             assert i not in url